+++ deeprank_gnn/add_self_loop.py
@@ -1,0 +1,16 @@
+import torch
+
+
+def add_self_loops_wattr(edge_index, edge_attr, num_nodes=None):
+    num_nodes = maybe_num_nodes(edge_index, num_nodes)
+
+    dtype, device = edge_index.dtype, edge_index.device
+    loop = torch.arange(0, num_nodes, dtype=dtype, device=device)
+    loop = loop.unsqueeze(0).repeat(2, 1)
+    edge_index = torch.cat([edge_index, loop], dim=1)
+
+    dtype, device = edge_attr.dtype, edge_attr.device
+    loop = torch.ones(num_nodes, dtype=dtype, device=device)
+    edge_attr = torch.cat([edge_attr, loop])
+
+    return edge_index, edge_attr
