+++ deeprank_gnn/sgat_conv.py
@@ -16,21 +16,6 @@ from torch_geometric.nn import max_pool_x
 from .community_pooling import get_preloaded_cluster, community_pooling
 
 
-def add_self_loops_wattr(edge_index, edge_attr, num_nodes=None):
-    num_nodes = maybe_num_nodes(edge_index, num_nodes)
-
-    dtype, device = edge_index.dtype, edge_index.device
-    loop = torch.arange(0, num_nodes, dtype=dtype, device=device)
-    loop = loop.unsqueeze(0).repeat(2, 1)
-    edge_index = torch.cat([edge_index, loop], dim=1)
-
-    dtype, device = edge_attr.dtype, edge_attr.device
-    loop = torch.ones(num_nodes, dtype=dtype, device=device)
-    edge_attr = torch.cat([edge_attr, loop])
-
-    return edge_index, edge_attr
-
-
 class sGraphAttention(torch.nn.Module):
 
     """
@@ -110,5 +95,3 @@ class sGraphAttention(torch.nn.Module):
         return '{}({}, {})'.format(self.__class__.__name__,
                                    self.in_channels,
                                    self.out_channels)
-
-
